feat: add initial root package (#1)
- Update repository title and remove intial description in `README.md` -
we'll add a proper description later.
- Add initial root package `dataorc` - this is currently an empty
package that will be used to test publishing to PyPI.
- Add initial Release Please configuration in
`release-please-config.json` and a release workflow
`.github/workflows/release.yml`.

Please note that the release workflow currently references a reusable
workflow from the main branch of
[equinor/ops-actions](https://github.com/equinor/ops-actions). The
version of that reusable workflow that we need to publish to PyPI has
not yet been released - we'll create and pin to a release tag once we've
verified the functionality of the workflow in this repository.