chore: configure root project (#30)
Configure the root `dataorc` project so that installing the `dataorc`
package will install all sub-packages in the `packages` directory. The
goal is for the `dataorc` package to function as an aggregator / meta
package for all sub-packages in this repository.

`pyproject.toml`:
- Configure the uv build backend, promoting consistent use of uv as our
main Python tool of choice.
- Add a dependency on the `dataorc-utils` package.
- Add the `[tool.uv.workspace]` table to configure a local uv workspace
for simplified dev environment setup - installing the root project will
install dependencies for all projects in the workspace.
- Add the `[tool.uv.sources]` table to configure a local dependency on
the `dataorc-utils` project in the workspace - during development, the
`dataorc-utils` package will be installed from the workspace project
instead of from PyPI.

`.github/dependabot.yml`:
- Configure Dependabot to keep dependencies in the root `pyproject.toml`
file up-to-date - will ensure that the root package always includes the
latest version of each package in the `packages` directory.

`packages/dataorc-utils/pyproject.toml`:
- Configure the `dataorc-utils` project to use the uv build backend for
the sake of consistency within this repository.
- Add empty `README.md` for the `dataorc-utils` project (uv requires all
projects in a workspace to have a README).

`scripts/bootstrap-package.sh`:
- Update the package bootstrap script to configure new projects to use
the uv build backend by default.

---------

Co-authored-by: TordAreStromsnes <toarst@equinor.com>